requirements.txt
fastapi
langchain
langchain-openai
openai
uvicorn
jinja2
python-dotenv